requirements.txt
requests==2.31.0
PyGithub==1.59.0
python-dotenv==1.0.0
Flask==2.2.5
Werkzeug==2.2.3
GitPython==3.1.31
Flask-Limiter==3.3.1
